refactor: getTopologySortedList 메서드를 별도의 외장 함수 형태로 추출
