+++ client/src/core/todo/todoList.ts
@@ -238,65 +238,4 @@ export class TodoList {
   async getTodoById(id: string): Promise<PlainTodo | undefined> {
     return this.todoList.find((el) => el.id === id)?.toPlain();
   }
-
-  async getTopologySortedList(filter: (todo: Todo) => boolean): Promise<Map<string, DiagramTodo>> {
-    const cloneTodoList = new Map<string, Todo>(
-      (await this.getSortedListWithFilter(filter, [])).map((el) => [el.id, new Todo(el)]),
-    );
-    const resultTodoList = new Map<string, DiagramTodo>(
-      (await this.getSortedListWithFilter(filter, [])).map((el) => [el.id, { depth: NaN, todo: new Todo(el) }]),
-    );
-
-    const updateDepth = (id: string, depth: number): void => {
-      const target = resultTodoList.get(id);
-      if (target !== undefined) {
-        target.depth = depth;
-      }
-    };
-
-    const forwardQueue = new Queue(
-      this.todoList
-        .filter((el) => filter(el) && this.checkPrev(el) && el.state !== 'DONE')
-        .map((el) => ({ depth: 0, id: el.id })),
-    );
-    while (!forwardQueue.isEmpty()) {
-      const target = forwardQueue.pop();
-      console.log(target);
-      updateDepth(target.id, target.depth);
-
-      const todo = cloneTodoList.get(target.id);
-      if (todo === undefined) continue;
-      [...todo.next].forEach((nextId) => {
-        const nextTodo = cloneTodoList.get(nextId);
-        if (nextTodo === undefined) return;
-        nextTodo.prev.delete(target.id);
-        if (nextTodo.prev.size === 0) {
-          forwardQueue.push({ depth: target.depth + 1, id: nextId });
-        }
-      });
-    }
-
-    const backwardQueue = new Queue(
-      this.todoList.filter((el) => this.checkPrev(el) && el.state !== 'DONE').map((el) => ({ depth: 0, id: el.id })),
-    );
-
-    while (!backwardQueue.isEmpty()) {
-      const target = backwardQueue.pop();
-      updateDepth(target.id, target.depth);
-
-      const todo = cloneTodoList.get(target.id);
-      if (todo === undefined) continue;
-      [...todo.prev].forEach((prevId) => {
-        const prevTodo = cloneTodoList.get(prevId);
-        if (prevTodo === undefined) return;
-        prevTodo.next.delete(target.id);
-        if (prevTodo.next.size === 0) {
-          backwardQueue.push({ depth: target.depth - 1, id: prevId });
-        }
-      });
-    }
-
-    console.log(resultTodoList);
-    return resultTodoList;
-  }
 }
